user auth and routes + task routes

+++ backend/routes/tasks.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from extensions import db
 from models import Task
 
@@ -11,13 +12,14 @@ def list_tasks():
     return jsonify([task.to_dict() for task in tasks])
 
 @tasks_bp.post("/")
+@jwt_required()
 def create_task():
-    # need authentication to get user id
-    # user_id = 
+    user_id = get_jwt_identity() 
+
     data = request.get_json()
 
     new_task = Task(
-        #creator_id=user_id
+        creator_id=user_id,
         title=data['title'],
         zipcode=data['zipcode'],
         description=data['description'],
@@ -41,19 +43,28 @@ def get_task(task_id):
     return jsonify(task.to_dict())
 
 @tasks_bp.delete('/<int:task_id>')
+@jwt_required()
 def delete_task(task_id):
-    #need auth
+    user_id = get_jwt_identity()
     task = Task.query.get_or_404(task_id)
+
+    if str(task.creator_id) != user_id:
+      return jsonify({'msg': 'only creator can delete this task'})
+
     db.session.delete(task)
     db.session.commit()
 
     return jsonify({'message': 'Task deleted'})
 
 @tasks_bp.put('/<int:task_id>')
+@jwt_required()
 def update_task(task_id):
-    #need auth to check if logged in user created task
+    user_id = get_jwt_identity()
     task = Task.query.get_or_404(task_id)
     data = request.get_json()
+
+    if str(task.creator_id) != user_id:
+      return jsonify({'msg': 'only creator can update this task'})
 
     if 'title' in data:
         task.title = data['title']
@@ -67,18 +78,27 @@ def update_task(task_id):
     return jsonify(task.to_dict())
 
 @tasks_bp.post('/<int:task_id>/accept')
+@jwt_required()
 def accept_task(task_id):
-    # Need authentication to complete this properly
+    user_id = get_jwt_identity()
     task =  Task.query.get_or_404(task_id)
-    task.status = 'in_progress'
+
+    if user_id == str(task.creator_id):
+        return jsonify({"msg": "user cannot accept their own task"})
+
+    task.accept(user_id)
     db.session.commit()
 
     return jsonify(task.to_dict())
 
 @tasks_bp.post('/<int:task_id>/complete')
+@jwt_required()
 def mark_task_complete(task_id):
-    #need auth
+    user_id = get_jwt_identity()
     task = Task.query.get_or_404(task_id)
+
+    if str(task.creator_id) != user_id:
+      return jsonify({'msg': 'only creator can mark task as completed'})
 
     task.mark_complete()
     db.session.commit()
